+++ packages/map/src/overlay/index.ts
@@ -1,3 +1,4 @@
-export * from './markers'
 export * from './polyline'
 export * from './polygon'
+export * from './circle-marker'
+export * from './bubble'
