--- packages/map/src/overlay/markers/index.ts
@@ -1,2 +1,0 @@
-export * from './circle-marker'
-export * from './marker-base'
